Validate NotNull fields automatically on persisting
This commit removes the usage of lombok's NonNull annotation which is
triggered on objects instantiation to
javax.validation.constraints.NotNull which will be used by a
MongoEventListener to validate objects that are about to be persisted.

This will be mainly useful for situations when it is cleaner or
necessary to build an object on different stages, for example,  a method
that adds a username to the user and another method that adds their
challenges. Another benefit this brings us is when we need to keep
subset of fields of a struct to do some operations on them; instead of
introducing a new struct we can reuse the existing struct without
running into null exceptions for the fields we don't need. But note that
we are still keeping the validations needed but only when the object is
about to be saved.